pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "watchdog>=6.0.0",
    "supabase>=2.7.4",
]
